hostapd-extn: Add support for compiling extensions with QCN_EXTN disabled.
This fix is added for use case where the CONFIG_PACKAGE_QCN_EXTN  is
enabled in the qsdk config, and extension code is applied to hostapd
build directory, but the CONFIG_QCN_EXTN flag is disabled while
compiling the code.

Add inline function definitions for required extension APIs in the
header files when the CONFIG_QCN_EXTN macro is not defined. This
avoids the linker error of undefined function.

Fixes: Ib5cb03e58187b1e688a484a3666a69c3c1558d46 (hostapd-extn: Add support to compile extensions for hostapd)

Signed-off-by: Sri Vidya Gunturi <sgunturi@qti.qualcomm.com>

Change-Id: Id3157a12f8f47b76e15eb9f52d8e3401e4f611c7